Consolidate parsing 'raw' EC keys
